move context_mgr to a submodule in identity
